bpf: move all unit tests to bpf/tests/
Place them into its own directory instead of directly into lib/
code. Adding more into lib/ would convolute it too much. Also
add a Wno-builtin-declaration-mismatch for gcc.

Tested via: # make unit-tests TESTPKGS=
